Moved index.php and bootstrap.php to separate folders.
