+++ README.md
@@ -4,4 +4,4 @@
 Steps to run the app:
 - In application root folder run the PHP built-in server with following command:
     php -S localhost:1991
-- Open localhost:1991 in your browser.
+- Open localhost:1991/public in your browser.
